+++ NEWS.md
@@ -1,5 +1,9 @@
 # bstfun (development version)
 
+* Added new function `add_sparkline()`--a wrapper for `gtExtras::gt_sparkline()` that adds distributional figures in a new column of a 'tbl_summary' table.
+
+* Updated default argument value: `as_forest_plot(xlog = x$inputs$exponentiate)`.
+
 # bstfun 0.2.6
 
 * Deprecated `tbl_2way_summary()` in favor of `gtsummary::tbl_continous()`.